+++ routes/news.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 var Msg = require('../bin/messages');
-var jsonEngine = require('../bin/json-engine');
+var jsonEngine = require('../bin/engines/json');
 
 
 /* GET News listing. */
